Split files in storage package and add basic test
